Change in Package name and also ready for play store updates
