+++ app/src/debug/res/values/google_maps_api.xml
@@ -20,5 +20,5 @@
     Once you have your key (it starts with "AIza"), replace the "google_maps_key"
     string in this file.
     -->
-    <string name="google_maps_key" templateMergeStrategy="preserve" translatable="false">AIzaSyDymO1j4czrwt92Wvf4pgCG0MDn5TbI1Y0</string>
+    <string name="google_maps_key" templateMergeStrategy="preserve" translatable="false">AIzaSyDfUaxnRrWEVN3qrzRlH8PzQoDp6XEuI5o</string>
 </resources>
