+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.kchan.activitytracker">
+    package="com.acitivitytracker.kchan.activitytracker">
 
     <!--
          The ACCESS_COARSE/FINE_LOCATION permissions are not required to use
@@ -33,7 +33,7 @@
             android:name="com.google.android.geo.API_KEY"
             android:value="@string/google_maps_key" />
 
-        <activity android:name=".SigninActivity">
+        <activity android:name="com.acitivitytracker.kchan.activitytracker.SigninActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -41,7 +41,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MapsActivity"
+            android:name="com.acitivitytracker.kchan.activitytracker.MapsActivity"
             android:label="@string/title_activity_maps" />
         <!--
      Service that provides activity recognition data. Setting the android:exported attribute
@@ -50,14 +50,14 @@
         -->
 
         <receiver
-            android:name=".BackgroundService.LocationUpdateBroadcastReceiver"
+            android:name="com.acitivitytracker.kchan.activitytracker.BackgroundService.LocationUpdateBroadcastReceiver"
             android:enabled="true"
             android:exported="true" />
 
-        <activity android:name=".DetailsActivity" />
+        <activity android:name="com.acitivitytracker.kchan.activitytracker.DetailsActivity" />
 
         <receiver
-            android:name=".BackgroundService.ActivityBroadcastReceiver"
+            android:name="com.acitivitytracker.kchan.activitytracker.BackgroundService.ActivityBroadcastReceiver"
             android:enabled="true"
             android:exported="true"></receiver>
     </application>
