--- app/src/main/java/com/example/kchan/activitytracker/Database/UserInfo.java
@@ -1,56 +1,0 @@
-package com.example.kchan.activitytracker.Database;
-
-public class UserInfo {
-
-    private String name;
-    private String email;
-    private String dob;
-    private String height;
-    private String weight;
-
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getDob() {
-        return dob;
-    }
-
-    public void setDob(String dob) {
-        this.dob = dob;
-    }
-
-    public String getHeight() {
-        return height;
-    }
-
-    public void setHeight(String height) {
-        this.height = height;
-    }
-
-    public String getWeight() {
-        return weight;
-    }
-
-    public void setWeight(String weight) {
-        this.weight = weight;
-    }
-
-    public UserInfo()
-    { }
-
-
-}
