--- app/src/main/java/com/example/kchan/activitytracker/GoogleSignInClientValue.java
@@ -1,32 +1,0 @@
-package com.example.kchan.activitytracker;
-
-import android.content.Context;
-
-import com.google.android.gms.auth.api.signin.GoogleSignIn;
-import com.google.android.gms.auth.api.signin.GoogleSignInClient;
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
-
-public class GoogleSignInClientValue {
-    // Build a GoogleSignInClient with the options specified by gso.
-    public GoogleSignInClient mGoogleSignInClient;
-    // Configure sign-in to request the user's ID, email address, and basic
-    // profile. ID and basic profile are included in DEFAULT_SIGN_IN.
-    GoogleSignInOptions gso ;
-
-    public GoogleSignInClientValue() {
-
-    }
-
-    public GoogleSignInClientValue(Context context) {
-        this.gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestIdToken(context.getString(R.string.default_web_client_id))
-                .requestEmail()
-                .build();
-        this.mGoogleSignInClient = GoogleSignIn.getClient(context, gso);
-    }
-
-    public GoogleSignInClient getInstance() {
-        return mGoogleSignInClient;
-    }
-
-}
