--- app/src/main/java/com/example/kchan/activitytracker/ResultHelper.java
@@ -1,165 +1,0 @@
-package com.example.kchan.activitytracker;
-
-import android.app.Notification;
-import android.app.NotificationChannel;
-import android.app.NotificationManager;
-import android.app.PendingIntent;
-import android.app.TaskStackBuilder;
-import android.content.Context;
-import android.content.Intent;
-import android.graphics.Color;
-import android.location.Location;
-import android.preference.PreferenceManager;
-
-import java.text.DateFormat;
-import java.util.Date;
-import java.util.List;
-
-public class ResultHelper {
-
-    final static String KEY_LOCATION_UPDATES_RESULT = "location-update-result";
-
-    final private static String PRIMARY_CHANNEL = "default";
-
-
-    private Context mContext;
-    private List<Location> mLocations;
-    private NotificationManager mNotificationManager;
-
-    public ResultHelper(Context context, List<Location> locations) {
-        mContext = context;
-        mLocations = locations;
-
-        NotificationChannel channel = new NotificationChannel(PRIMARY_CHANNEL,
-                context.getString(R.string.default_channel), NotificationManager.IMPORTANCE_DEFAULT);
-        channel.setLightColor(Color.GREEN);
-        channel.setLockscreenVisibility(Notification.VISIBILITY_PRIVATE);
-        getNotificationManager().createNotificationChannel(channel);
-    }
-
-    public ResultHelper(Context context) {
-        mContext = context;
-
-        NotificationChannel channel = new NotificationChannel(PRIMARY_CHANNEL,
-                context.getString(R.string.default_channel), NotificationManager.IMPORTANCE_DEFAULT);
-        channel.setLightColor(Color.GREEN);
-        channel.setLockscreenVisibility(Notification.VISIBILITY_PRIVATE);
-        getNotificationManager().createNotificationChannel(channel);
-    }
-
-    /**
-     * Returns the title for reporting about a list of {@link Location} objects.
-     */
-    private String getLocationResultTitle() {
-        String numLocationsReported = mContext.getResources().getQuantityString(
-                R.plurals.num_locations_reported, mLocations.size(), mLocations.size());
-        return numLocationsReported + ": " + DateFormat.getDateTimeInstance().format(new Date());
-    }
-
-    private String getLocationResultText() {
-        if (mLocations.isEmpty()) {
-            return mContext.getString(R.string.unknown_location);
-        }
-        StringBuilder sb = new StringBuilder();
-        for (Location location : mLocations) {
-            sb.append("(");
-            sb.append(location.getLatitude());
-            sb.append(", ");
-            sb.append(location.getLongitude());
-            sb.append(")");
-            sb.append("\n");
-        }
-        return sb.toString();
-    }
-
-    /**
-     * Saves location result as a string to {@link android.content.SharedPreferences}.
-     */
-    public void saveResults() {
-        PreferenceManager.getDefaultSharedPreferences(mContext)
-                .edit()
-                .putString(KEY_LOCATION_UPDATES_RESULT, getLocationResultTitle() + "\n" +
-                        getLocationResultText())
-                .apply();
-    }
-
-    /**
-     * Fetches location results from {@link android.content.SharedPreferences}.
-     */
-    public static String getSavedLocationResult(Context context) {
-        return PreferenceManager.getDefaultSharedPreferences(context)
-                .getString(KEY_LOCATION_UPDATES_RESULT, "");
-    }
-
-    /**
-     * Get the notification mNotificationManager.
-     * <p>
-     * Utility method as this helper works with it a lot.
-     *
-     * @return The system service NotificationManager
-     */
-    private NotificationManager getNotificationManager() {
-        if (mNotificationManager == null) {
-            mNotificationManager = (NotificationManager) mContext.getSystemService(
-                    Context.NOTIFICATION_SERVICE);
-        }
-        return mNotificationManager;
-    }
-
-    /**
-     * Displays a notification with the location results.
-     */
-    public void showNotification() {
-        Intent notificationIntent = new Intent(mContext, MapsActivity.class);
-
-        // Construct a task stack.
-        TaskStackBuilder stackBuilder = TaskStackBuilder.create(mContext);
-
-        // Add the main Activity to the task stack as the parent.
-        stackBuilder.addParentStack(MapsActivity.class);
-
-        // Push the content Intent onto the stack.
-        stackBuilder.addNextIntent(notificationIntent);
-
-        // Get a PendingIntent containing the entire back stack.
-        PendingIntent notificationPendingIntent =
-                stackBuilder.getPendingIntent(0, PendingIntent.FLAG_UPDATE_CURRENT);
-
-        Notification.Builder notificationBuilder = new Notification.Builder(mContext,
-                PRIMARY_CHANNEL)
-                .setContentTitle(getLocationResultTitle())
-                .setContentText(getLocationResultText())
-                .setSmallIcon(R.mipmap.ic_launcher)
-                .setAutoCancel(true)
-                .setContentIntent(notificationPendingIntent);
-
-        getNotificationManager().notify(0, notificationBuilder.build());
-    }
-
-    public void showNotificationActivity(String update) {
-        Intent notificationIntent = new Intent(mContext, MapsActivity.class);
-
-        // Construct a task stack.
-        TaskStackBuilder stackBuilder = TaskStackBuilder.create(mContext);
-
-        // Add the main Activity to the task stack as the parent.
-        stackBuilder.addParentStack(MapsActivity.class);
-
-        // Push the content Intent onto the stack.
-        stackBuilder.addNextIntent(notificationIntent);
-
-        // Get a PendingIntent containing the entire back stack.
-        PendingIntent notificationPendingIntent =
-                stackBuilder.getPendingIntent(1, PendingIntent.FLAG_UPDATE_CURRENT);
-
-        Notification.Builder notificationBuilder = new Notification.Builder(mContext,
-                PRIMARY_CHANNEL)
-                .setContentTitle("Activity Recognition On")
-                .setContentText(update)
-                .setSmallIcon(R.mipmap.ic_launcher)
-                .setAutoCancel(true)
-                .setContentIntent(notificationPendingIntent);
-
-        getNotificationManager().notify(1, notificationBuilder.build());
-    }
-}
