+++ app/src/release/res/values/google_maps_api.xml
@@ -16,5 +16,5 @@
     Once you have your key (it starts with "AIza"), replace the "google_maps_key"
     string in this file.
     -->
-    <string name="google_maps_key" templateMergeStrategy="preserve" translatable="false">YOUR_KEY_HERE</string>
+    <string name="google_maps_key" templateMergeStrategy="preserve" translatable="false">AIzaSyDfUaxnRrWEVN3qrzRlH8PzQoDp6XEuI5o</string>
 </resources>
