--- app/src/test/java/com/example/kchan/activitytracker/ExampleUnitTest.java
@@ -1,17 +1,0 @@
-package com.example.kchan.activitytracker;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() {
-        assertEquals(4, 2 + 2);
-    }
-}
